LPS-57513 Move classes out of internal package and export
