+++ modules/portal-cache/portal-cache-ehcache/bnd.bnd
@@ -1,5 +1,7 @@
 Bundle-Name: Liferay Portal Cache Ehcache
 Bundle-SymbolicName: com.liferay.portal.cache.ehcache
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.portal.cache.ehcache
+Export-Package: \
+	com.liferay.portal.cache.ehcache,\
+	com.liferay.portal.cache.ehcache.event
 Include-Resource: classes
