+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/EhcacheConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.internal;
+package com.liferay.portal.cache.ehcache;
 
 /**
  * @author Tina Tian
