+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/event/EhcachePortalCacheManagerListenerAdapter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.internal.event;
+package com.liferay.portal.cache.ehcache.event;
 
 import com.liferay.portal.kernel.cache.PortalCacheException;
 import com.liferay.portal.kernel.cache.PortalCacheManagerListener;
