+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/EhcacheConfigurationHelperUtil.java
@@ -14,6 +14,7 @@
 
 package com.liferay.portal.cache.ehcache.internal;
 
+import com.liferay.portal.cache.ehcache.EhcacheConstants;
 import com.liferay.portal.kernel.cache.PortalCacheListenerScope;
 import com.liferay.portal.kernel.cache.PortalCacheReplicator;
 import com.liferay.portal.kernel.cache.cluster.ClusterLinkCallbackFactory;
