+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/bootstrap/EhcachePortalCacheBootstrapLoaderAdapter.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.cache.ehcache.internal.bootstrap;
 
-import com.liferay.portal.cache.ehcache.internal.EhcacheUnwrapUtil;
+import com.liferay.portal.cache.ehcache.EhcacheUnwrapUtil;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheBootstrapLoader;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
