+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/distribution/EhcachePortalCacheReplicatorAdapter.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.cache.ehcache.internal.distribution;
 
-import com.liferay.portal.cache.ehcache.internal.event.EhcachePortalCacheListenerAdapter;
+import com.liferay.portal.cache.ehcache.event.EhcachePortalCacheListenerAdapter;
 import com.liferay.portal.kernel.cache.PortalCacheReplicator;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
